requirements.txt
python-docx==1.1.0
openpyxl==3.1.2
pyyaml==6.0.1